examples/local_docs/requirements.txt
haystack-ai
ollama-haystack
pgvector-haystack
datasets>=2.6.1
pypdf
python-docx
markdown-it-py
mdit-plain
